requirements.txt
adbutils==2.7.2
allure-pytest==2.13.5
allure-python-commons==2.13.5
amqp==5.2.0
apkutils2==1.0.0
Appium-Python-Client==2.10.2
async-timeout==4.0.3
attrs==23.2.0
backports.zoneinfo==0.2.1
billiard==4.2.0
blinker==1.8.2
celery==5.4.0
certifi==2024.2.2
charset-normalizer==3.3.2
cigam==0.0.3
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
configparser==7.0.0
decorator==5.1.1
deprecation==2.1.0
exceptiongroup==1.2.1
ffmpeg-python==0.2.0
filelock==3.14.0
Flask==3.0.3
Flask-Celery==2.4.3
Flask-Script==2.0.6
fsspec==2024.5.0
future==1.0.0
h11==0.14.0
idna==3.7
importlib_metadata==8.0.0
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
kombu==5.3.7
lpips==0.1.4
lxml==5.2.2
MarkupSafe==2.1.5
mpmath==1.3.0
natsort==8.4.0
networkx==3.1
numpy==1.24.4
outcome==1.3.0.post0
packaging==24.0
pillow==10.3.0
pluggy==1.5.0
prompt_toolkit==3.0.47
py==1.11.0
pyelftools==0.31
PyMySQL==1.1.1
PySocks==1.7.1
pytest==8.2.1
python-dateutil==2.9.0.post0
redis==5.0.7
requests==2.32.3
retry==0.9.2
scipy==1.10.1
selenium==4.8.2
shutils==0.1.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
sympy==1.12
tomli==2.0.1
torch==2.3.0
torchvision==0.18.0
tqdm==4.66.4
trio==0.25.1
trio-websocket==0.11.1
typing_extensions==4.12.0
tzdata==2024.1
uiautomator2==3.1.0
urllib3==1.26.18
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.0.3
wsproto==1.2.0
xmltodict==0.13.0
zipp==3.19.2
